move tests in tests directory
